// src/api/gen/models/AthleteExerciseHistoryResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AthleteExerciseHistoryEntry } from "./AthleteExerciseHistoryEntry";
export type AthleteExerciseHistoryResponse = {
  exerciseId?: string | null;
  entries: Array<AthleteExerciseHistoryEntry>;
};