// src/api/gen/models/ExerciseLastPerformanceResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LastPerformanceSet } from "./LastPerformanceSet";
export type ExerciseLastPerformanceResponse = {
  exerciseId?: string | null;
  workoutInstanceId?: string | null;
  performedAt?: string | null;
  sets: Array<LastPerformanceSet>;
};