// src/api/gen/models/SetExecution.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SetExecution = {
  id: string;
  setNumber: number;
  weight?: number;
  reps?: number;
  rpe?: number;
  isCompleted: boolean;
  restSecondsActual?: number;
  createdAt?: string;
  updatedAt?: string;
};