// src/api/gen/models/WorkoutLastCompletionResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WorkoutLastCompletionResponse = {
  workoutId: string;
  completedAt: string;
  durationSeconds: number;
  totalSets: number;
  completedSets: number;
  totalReps: number;
  volume: number;
};