// src/api/gen/models/WorkoutScheduleResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WorkoutScheduleItem } from "./WorkoutScheduleItem";
export type WorkoutScheduleResponse = {
  enrollmentId: string;
  items: Array<WorkoutScheduleItem>;
};